[BIG] Avoid this kind of big changes in the future
This commit completes the main goal of the app: measuring the speed of
reading (words per minute) and analyzing the comprehension of the text,
by using a quiz.

Unfortunately, it got ou of control, since I lost track of the changes
after a while, and could not separate in smaller alterations.